Added sorting by field to song display
The simplest approach was to store all numerical data as Integer and Long. Now, that data will be null (instead of -1) if it doesn't exist. The result is easier to program, but it worries the typesafe programmer inside of me. (Plus, it takes up more space to store all those Integer references than primitive ints.)