Added XML file saving
The machinery for writing XML files already existed, but this commit adds a chooser GUI as well as saving the current file.

This commit also adds a hack to put a slash ("/") at the end of a file name if that file name appears to be a directory (because it doesn't have an extension).