Fixed display of adding/removing files
Now, the convention is to set the playlist to the current playlist when a file is added or removed. Doing so refreshes the files in the JTable.

This commit also contains a bug fix in which a file was removed from a list of files created for display purposes rather than the underlying Map in the Library.